# Optimize encode_length