@ojwb 2019-05-02
That sounds like a nice improvement, but what mix of values are you benchmarking on here?

I'd expect values which fit in a byte to dominate in practice because term lengths should always actually fit in a byte and most wdf values probably will too, but for example timing something like `for (unsigned v = 0; v < UINT_MAX; ++v) encode_length(v);` will only test 255 values which fit in a byte and more than 16 million times as many which don't.

If you're able to, sharing the benchmark code would be helpful.  It would allow checking that we're not optimising this for one compiler at the expense of others, and allow others to more easily optimise further if they want to.

I've also wondered if the remote backend should just use `pack_uint()` as the on-disk formats mostly do.  This uses a slightly simpler encoding, which is 2 bytes instead of one for values 128-254 but is a byte shorter for some larger values.  So it'd probably increase the message sizes a little, but probably reduce CPU.  If we had a benchmark based on some representative data we could check how the size and speed compared.