@Kronuz 2019-05-02
I’m pretty sure I added the benchmarking code in the ticket i opened about this...

Here’s the ticket with the benchmarking code: https://trac.xapian.org/ticket/782