@ojwb 2019-05-02
Ah sorry, i missed that there was a ticket for this.  Thanks for the link.

So it seems the benchmark uses a logarithmically skewed set of values, which probably isn't too far from reality (on a typical platform ~1/8 of those values will fit in a byte, which is likely too low, but probably only a few times too low).