@ojwb 2019-06-08
After much further analysis and discussion in https://trac.xapian.org/ticket/782 we've decided to switch to using `pack_uint()` and `pack_unit_last()` instead, which are faster still, already used extensively in the library, and should typically result in smaller serialisations.

So I'm going to close this, but thanks for the patch, and for raising the issue and setting up the benchmarks of it.