@ojwb 2019-06-18
db790e9e12bb9b3ebeaf916ac0acdea9a7ab0dd1 implements the change to use `pack_uint()`, etc.